Add updated relationships between tables
