+++ service/models/__init__.py
@@ -3,6 +3,5 @@ Models for Payments service
 
 """
 
-from .persistent_base import db, DataValidationError
-from .payment_method import PaymentMethod, PaymentType
+from .payment_method import PaymentMethod, PaymentMethodType, DataValidationError, db
 from .credit_card import CreditCard, FieldValidationError
