+++ service/models/payment_method.py
@@ -4,77 +4,115 @@ Models for PaymentMethod
 All of the models are stored in this module
 """
 
+import logging
 from enum import Enum
-from .persistent_base import PersistentBase, DataValidationError, db, logger
+from abc import abstractmethod
+from flask_sqlalchemy import SQLAlchemy
+
+logger = logging.getLogger("flask.app")
+
+db = SQLAlchemy()
 
 
-class PaymentType(Enum):
+class DataValidationError(Exception):
+    """Used for an data validation errors when deserializing"""
+
+
+class PaymentMethodType(Enum):
     """Enumeration of valid payment types"""
 
-    UNKNOWN = 0
     CREDIT_CARD = 1
     PAYPAL = 2
 
 
-class PaymentMethod(db.Model, PersistentBase):
-    """
-    Class that represents a PaymentMethod
-    """
+class PaymentMethod(db.Model):
+    """Class that represents Payment Method resource"""
 
     ##################################################
-    # Table Schema
+    # TABLE SCHEMA
     ##################################################
+
     id = db.Column(db.Integer, autoincrement=True, primary_key=True)
     name = db.Column(db.String(63), nullable=False)
-    payment_type = db.Column(
-        db.Enum(PaymentType), nullable=False, server_default=(PaymentType.UNKNOWN.name)
-    )
-    payment_type_id = db.Column(db.Integer)
     user_id = db.Column(db.Integer, nullable=False)
+    type = db.Column(db.Enum(PaymentMethodType), nullable=False)
+
+    # https://docs.sqlalchemy.org/en/20/orm/inheritance.html
+    #
+    # We are utilizing the SQLAlchemy inheritance hierarchy with polymorphic identities
+    # This allows us to query all payment methods from PaymentMethod class
+    __mapper_args__ = {
+        "polymorphic_on": "type",
+    }
 
     def __repr__(self):
         return f"<PaymentMethod {self.name} id=[{self.id}]>"
 
-    def serialize(self):
-        """Serializes a PaymentMethod into a dictionary"""
-        return {
-            "id": self.id,
-            "name": self.name,
-            "payment_type": self.payment_type.name,
-            "payment_type_id": self.payment_type_id,
-            "user_id": self.user_id,
-        }
+    @abstractmethod
+    def serialize(self) -> dict:
+        """Convert an object into a dictionary"""
 
-    def deserialize(self, data):
-        """
-        Deserializes a PaymentMethod from a dictionary
+    @abstractmethod
+    def deserialize(self, data: dict) -> None:
+        """Convert a dictionary into an object"""
 
-        Args:
-            data (dict): A dictionary containing the resource data
+    def create(self) -> None:
         """
+        Creates a PaymentMethod to the database
+        """
+        logger.info("Creating %s", self)
+        # id must be none to generate next primary key
+        self.id = None
         try:
-            self.name = data["name"]
-            self.payment_type_id = data["payment_type_id"]
-            self.payment_type = getattr(
-                PaymentType, data["payment_type"]
-            )  # create enum from string
-            self.user_id = data["user_id"]
-        except AttributeError as error:
-            raise DataValidationError("Invalid attribute: " + error.args[0]) from error
-        except KeyError as error:
-            raise DataValidationError(
-                "Invalid PaymentMethod: missing " + error.args[0]
-            ) from error
-        except TypeError as error:
-            raise DataValidationError(
-                "Invalid PaymentMethod: body of request contained bad or no data "
-                + str(error)
-            ) from error
-        return self
+            db.session.add(self)
+            db.session.commit()
+        except Exception as e:
+            db.session.rollback()
+            logger.error("Error creating record: %s", self)
+            raise DataValidationError(e) from e
+
+    def update(self) -> None:
+        """
+        Updates a PaymentMethod to the database
+        """
+        logger.info("Updating %s", self)
+        if not self.id:
+            raise DataValidationError("Update called with empty ID field")
+        try:
+            db.session.commit()
+        except Exception as e:
+            db.session.rollback()
+            logger.error("Error updating record: %s", self)
+            raise DataValidationError(e) from e
+
+    def delete(self) -> None:
+        """Removes a PaymentMethod from the data store"""
+        logger.info("Deleting %s", self)
+        try:
+            db.session.delete(self)
+            db.session.commit()
+        except Exception as e:
+            db.session.rollback()
+            logger.error("Error deleting record: %s", self)
+            raise DataValidationError(e) from e
 
     ##################################################
     # CLASS METHODS
     ##################################################
+
+    @classmethod
+    def all(cls):
+        """Returns all of the records in the database"""
+        logger.info("Processing all records")
+        # pylint: disable=no-member
+        return cls.query.all()
+
+    @classmethod
+    def find(cls, by_id):
+        """Finds a record by it's ID"""
+        logger.info("Processing lookup for id %s ...", by_id)
+        # pylint: disable=no-member
+        return cls.query.session.get(cls, by_id)
 
     @classmethod
     def find_by_name(cls, name):
